+++ web/src/components/travel-authorizations/details-edit-form-card/MultiDestinationStopsSection.vue
@@ -53,7 +53,7 @@
         cols="12"
         md="2"
       >
-        <TimePicker
+        <TimePickerLegacy
           :value="firstStop.departureTime"
           label="Time (24 hour)"
           :field-options="{
@@ -150,7 +150,7 @@
         cols="12"
         md="2"
       >
-        <TimePicker
+        <TimePickerLegacy
           :value="secondStop.departureTime"
           label="Time (24 hour)"
           :field-options="{
@@ -261,7 +261,7 @@
           cols="12"
           md="2"
         >
-          <TimePicker
+          <TimePickerLegacy
             :value="stops[index + 2].departureTime"
             label="Time (24 hour)"
             :field-options="{
@@ -320,7 +320,7 @@ import { required, greaterThanOrEqualToDate } from "@/utils/validators"
 import AccommodationTypeSelect from "@/modules/travel-authorizations/components/AccommodationTypeSelect.vue"
 import DatePicker from "@/components/common/DatePicker.vue"
 import LocationsAutocomplete from "@/components/locations/LocationsAutocomplete.vue"
-import TimePicker from "@/components/common/TimePicker.vue"
+import TimePickerLegacy from "@/components/common/TimePickerLegacy.vue"
 import TravelMethodSelect from "@/modules/travel-authorizations/components/TravelMethodSelect.vue"
 
 const props = defineProps({
