+++ web/src/components/travel-segments/TravelSegmentsAttributesMultiCityTripSection.vue
@@ -53,7 +53,7 @@
         cols="12"
         md="3"
       >
-        <TimePicker
+        <TimePickerLegacy
           v-model="firstTravelSegmentAttributes.departureTime"
           label="Time (24 hour)"
           :field-options="{
@@ -204,7 +204,7 @@
           cols="12"
           md="3"
         >
-          <TimePicker
+          <TimePickerLegacy
             v-model="travelSegmentsAttributes[index].departureTime"
             label="Time (24 hour)"
             :field-options="{
@@ -301,7 +301,7 @@ import {
 } from "@/api/travel-segments-api"
 
 import DatePicker from "@/components/common/DatePicker.vue"
-import TimePicker from "@/components/common/TimePicker.vue"
+import TimePickerLegacy from "@/components/common/TimePickerLegacy.vue"
 import LocationsAutocomplete from "@/components/locations/LocationsAutocomplete.vue"
 import AccommodationTypeSelect from "@/components/travel-segments/AccommodationTypeSelect.vue"
 import TravelMethodSelect from "@/components/travel-segments/TravelMethodSelect.vue"
