+++ web/src/components/travel-segments/TravelSegmentsAttributesOneWayTripSection.vue
@@ -52,7 +52,7 @@
         cols="12"
         md="3"
       >
-        <TimePicker
+        <TimePickerLegacy
           v-model="departTravelSegmentAttributes.departureTime"
           label="Time (24 hour)"
           :field-options="{
@@ -138,7 +138,7 @@ import {
 } from "@/api/travel-segments-api"
 
 import DatePicker from "@/components/common/DatePicker.vue"
-import TimePicker from "@/components/common/TimePicker.vue"
+import TimePickerLegacy from "@/components/common/TimePickerLegacy.vue"
 import AccommodationTypeSelect from "@/components/travel-segments/AccommodationTypeSelect.vue"
 import LocationsAutocomplete from "@/components/locations/LocationsAutocomplete.vue"
 import TravelMethodSelect from "@/components/travel-segments/TravelMethodSelect.vue"
